# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API CashWarden",
        "description": "Toutes les requêtes API CashWarden",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/requisition": {
            "post": {
                "tags": [
                    "Requisition"
                ],
                "summary": "Permet de se connecter",
                "description": "Déconnecte l'utilisateur",
                "operationId": "postRequisition",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreRequisitionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "": {
                                                                "type": "string",
                                                                "example": "Merci de renseigner un identifiant de banque !"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "An example resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "lastname": {
                                                    "type": "string",
                                                    "example": "DUPONT"
                                                },
                                                "firstname": {
                                                    "type": "string",
                                                    "example": "Jean"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "dupont.jean@gmail.com"
                                                },
                                                "bank_id": {
                                                    "type": "string",
                                                    "example": "null"
                                                },
                                                "IdRequisition": {
                                                    "type": "string",
                                                    "example": "null"
                                                },
                                                "account_id": {
                                                    "type": "string",
                                                    "example": "null"
                                                },
                                                "hasBankAuthorization": {
                                                    "type": "boolean",
                                                    "example": "false"
                                                },
                                                "hasAccountChoices": {
                                                    "type": "boolean",
                                                    "example": "false"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "requisition_response": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "0e981e62-f803-4ff1-a775-eac2e6af6a2c"
                                                },
                                                "created": {
                                                    "type": "string",
                                                    "example": "2022-03-17T23:07:51.570255Z"
                                                },
                                                "redirect": {
                                                    "type": "string",
                                                    "example": "http://localhost:3000/bank-register/"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "CR"
                                                },
                                                "institution_id": {
                                                    "type": "string",
                                                    "example": "AGRICOLE_SUD_RHONE_ALPES_AGRIFRPPXXX"
                                                },
                                                "agreement": {
                                                    "type": "boolean",
                                                    "example": ""
                                                },
                                                "reference": {
                                                    "type": "boolean",
                                                    "example": "9f6233b3-e1dc-41b8-96f3-c4e423a74266"
                                                },
                                                "accounts": {
                                                    "type": "json",
                                                    "example": "[]"
                                                },
                                                "link": {
                                                    "type": "boolean",
                                                    "example": "https://ob.nordigen.com/psd2/start/9f6233b3-e1dc-41b8-96f3-c4e423a74266/AGRICOLE_SUD_RHONE_ALPES_AGRIFRPPXXX"
                                                },
                                                "ssn": {
                                                    "type": "boolean",
                                                    "example": "null"
                                                },
                                                "account_selection": {
                                                    "type": "boolean",
                                                    "example": "false"
                                                },
                                                "redirect_immediate": {
                                                    "type": "boolean",
                                                    "example": "false"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreRequisitionRequest": {
                "properties": {
                    "bank_id": {
                        "type": "string",
                        "example": "AGRICOLE_SUD_RHONE_ALPES_AGRIFRPPXXX"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}